feat [#119] 팀 생성 및 수정 페이지를 공통 컴포넌트 기반으로 수정, 팀 수정 페이지에 SSR 적용
